[wip] clients can connect to backend through WS
test is failing as the echo backend does not respect the
protocol (eg. it does not send Result objects) and the intermediate
functions in Player.hs have failing patterns so it breaks connection
with weird results.

TODO: Handle potentially invalid responses from the backend server in a more
sensible way