Adjust max PR spent on tech
The change to limit tech improvements to what's available had a deep
impact on Empires' chances of winning. I think this stems from the
fact it's more likely for robot to Pass when choosing tech because the
number of alternate choices is significantly smaller. Why Empires is
more affected is not clear though.

Changing the maximum spent from 5 to 4 provides a 8% increase of
Empires' wins, but it's still lower than it used to be :shrug: